Rename ps_installers to ps_loader; Pyodide uses mip.py for MIP installs
- ps_loader.install_pyodide() calls add_ons/mip for manifests/modules
  (url_base=page root for ./src/… manifest paths) and micropip for wheels
- Shared _install_manifests_and_modules() for both runtimes
- mip.py: restore Pyodide open_url HTTP; add optional url_base kwarg
- Unit test for url_base relative manifest resolution
