+++ web/pyscript/run-pyodide.html
@@ -78,7 +78,7 @@
 <script type="py" config="./pyodide.toml" output="log">
 import builtins
 from js import document
-import ps_installers
+import ps_loader
 
 def _log_print(*args, **kwargs):
     el = document.getElementById("log")
@@ -97,9 +97,9 @@ def _from_window(name, default=""):
 async def _run():
     if _from_window("__pydisplayLoaderDisabled"):
         return
-    modules = ps_installers.parse_names(_from_window("__pydisplayModules"))
-    manifests = ps_installers.parse_names(_from_window("__pydisplayManifests"))
-    wheels = ps_installers.parse_names(_from_window("__pydisplayWheels"))
+    modules = ps_loader.parse_names(_from_window("__pydisplayModules"))
+    manifests = ps_loader.parse_names(_from_window("__pydisplayManifests"))
+    wheels = ps_loader.parse_names(_from_window("__pydisplayWheels"))
     entry = _from_window("__pydisplayEntry")
     if not entry and modules:
         entry = modules[0]
@@ -108,7 +108,7 @@ async def _run():
     if not entry:
         print("Missing ?modules= and/or ?manifests= in the URL.")
         return
-    await ps_installers.install_pyodide(modules, manifests, wheels)
+    await ps_loader.install_pyodide(modules, manifests, wheels)
     __import__(entry)
 
 await _run()
